refactor(capsule): move capsule-list and capsule-paths to network
