+++ src/extensions/capsule/capsule.ts
@@ -18,16 +18,6 @@ export default class CapsuleFactory {
   ) {}
 
   /**
-   * create a new capsule from a component.
-   */
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  //   create(components: Component[], config?: CapsuleOptions) {
-  //     return this.builder.isolateComponents(
-  //       components.map(component => component.id.toString()),
-  //       config
-  //     );
-  //   }
-  /**
    * list all of the existing workspace capsules.
    */
   list(): ComponentCapsule[] {
